ai_library_bot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_library_bot"
version = "0.1.0"
description = "RAG-based Telegram bot for answering questions from uploaded books"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "openai>=1.35.0",
    "faiss-cpu>=1.7.4",
    "ebooklib>=0.18",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "aiocache>=0.12.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[tool.ruff]
# Конфигурация ruff (быстрый линтер)
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports allowed in __init__.py
"tests/*" = ["E501", "F811"]  # allow long lines and redefined names in tests

[tool.black]
# Конфигурация black (форматтер)
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
# Конфигурация mypy (проверка типов)
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_optional = true

# Игнорируем отсутствующие импорты для внешних библиотек
[[tool.mypy.overrides]]
module = [
    "telegram.*",
    "faiss.*",
    "ebooklib.*",
    "PyPDF2.*",
    "aiocache.*",
    "langchain.*",
    "langchain_openai.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
# Конфигурация pytest
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]

[tool.setuptools]
packages = ["src"]

